chore(sdk-testing): re-factor testing examples
- Refactor testing examples
- Remove InvokeAccount for testing cases
