+++ examples/src/parallel/parallel_first_successful.py
