+++ examples/test/step/test_step.py
@@ -3,7 +3,7 @@
 import pytest
 from aws_durable_execution_sdk_python.execution import InvocationStatus
 
-from src import step
+from src.step import step
 from test.conftest import deserialize_operation_payload
