+++ examples/test/step/test_steps_with_retry.py
@@ -4,7 +4,7 @@ import pytest
 from aws_durable_execution_sdk_python.execution import InvocationStatus
 from aws_durable_execution_sdk_python.lambda_service import OperationType
 
-from src import steps_with_retry
+from src.step import steps_with_retry
 from test.conftest import deserialize_operation_payload
